+++ packages/jest-jasmine2/src/__tests__/expectationResultFactory.test.ts
@@ -55,6 +55,18 @@ describe('expectationResultFactory', () => {
     expect(result.message).toEqual('Error: Expected `Pass`, received `Fail`.');
   });
 
+  it('returns the error name if the error message is empty', () => {
+    const options = {
+      actual: 'Fail',
+      error: new Error(),
+      expected: 'Pass',
+      matcherName: 'testMatcher',
+      passed: false,
+    };
+    const result = expectationResultFactory(options);
+    expect(result.message).toEqual('Error');
+  });
+
   it('returns the result if failed (with `error` as a string).', () => {
     const options = {
       actual: 'Fail',